docker

+++ test/docker/mkenv.sh
@@ -7,7 +7,9 @@ source $thispath/common.sh
 
 function init() {
   container_list=$thispath/container_list
-  rm $container_list
+  if [ -f $container_list ]; then
+    rm $container_list
+  fi
   cport=1179
 }
 
@@ -23,24 +25,26 @@ function make_nodes() {
   for ((i=0; i<$node_num; i++)); do
     cname="seris_n$i"
     cid=`run_container_1 $cname`
+    if [ "$cid" = "" ]; then
+      continue
+    fi
     cip=`docker inspect $cid | grep "\"IPAddress\"" | head -1 | awk -F : '{ print $2 }' | sed 's/[", ]//g'`
     echo "$cid,$cname,$cip" >> $container_list
   done
 }
 
 function configure_nodes() {
+  #for line in `cat $container_list`; do
+  for ((i=0; i<$node_num; i++)); do
+    #cid=`echo $line | awk -F , '{ print $1 }'`
+    #cname=`echo $line | awk -F , '{ print $2 }'`
+    #cip=`echo $line | awk -F , '{ print $3 }'`
 
-  for line in `cat $container_list`; do
-    cid=`echo $line | awk -F , '{ print $1 }'`
-    cname=`echo $line | awk -F , '{ print $2 }'`
-    cip=`echo $line | awk -F , '{ print $3 }'`
-
-    docker cp $SERIS_HOME $cid:/usr/local/
-
+    #docker cp $SERIS_HOME $cid:/usr/local/
     # meta/selfnode
-    rm -rf $thispath/meta
-    mkdir -p $thispath/meta/work
-    echo $cname,$cip,$cport > $thispath/meta/selfnode
+    #rm -rf $thispath/meta
+    #mkdir -p $thispath/meta
+    #echo $cname,$cip,$cport > $thispath/meta/selfnode
 
     # meta/nodelist
     friend_num=`get_friends_num $node_num`
@@ -61,11 +65,22 @@ function configure_nodes() {
     done
 
     # meta/env
-    echo "export SERIS_NODEID="$cname > $thispath/meta/env
-    echo "export SERIS_HOST="$cip >> $thispath/meta/env
-    echo "export SERIS_PORT="$cport >> $thispath/meta/env
+    #echo "export SERIS_NODEID="$cname > $thispath/meta/env
+    #echo "export SERIS_HOST="$cip >> $thispath/meta/env
+    #echo "export SERIS_PORT="$cport >> $thispath/meta/env
+    #docker cp $thispath/meta $cid:/usr/local/seris/
 
-    docker cp $thispath/meta $cid:/usr/local/seris/
+    #cname="seris_n$i"
+    #cid=`run_container_1`
+    #if [ "$cid" = "" ]; then
+    #  continue
+    #fi
+    #cip=`docker inspect $cid | grep "\"IPAddress\"" | head -1 | awk -F : '{ print $2 }' | sed 's/[", ]//g'`
+    #docker exec -it $SERIS_HOME/src/invoke.sh $SERIS_HOST $SERIS_PORT seris/addenv SERIS_NODEID $cid
+    #docker exec -it $SERIS_HOME/src/invoke.sh $SERIS_HOST $SERIS_PORT seris/addenv SERIS_HOST $cip
+    #docker exec -it $SERIS_HOME/src/invoke.sh $SERIS_HOST $SERIS_PORT seris/addenv SERIS_PORT $cport
+    #docker exec -it $SERIS_HOME/src/invoke.sh $SERIS_HOST $SERIS_PORT seris/addenv SERIS_LINKED_LIST $SERIS_LINKED_LIST
+    #echo "$cid,$cname,$cip" >> $container_list
   done
 }
